Fix bug where app would crash
Before this commit, the app would upload the teams to Firebase, notice
that someone uploaded something to Firebase, and then download the
teams from firebsae, at which point it would notice that the local
teams list had been changed, so it would reupload it to Firebase,
restarting the cycle until the app ran out of memory and froze.

Now, it only uploads to Firebase when the app starts or when a new
team is added, solving the bug.